forgot password

+++ backend/controllers/userController.js
@@ -108,7 +108,7 @@ export const RegisterUser = async (req, res, next) => {
 export const SendOtp = async (req, res, next) => {
   try {
     // get the phone_number from req.body
-    const { first_name = "", last_name = "", email = "" } = req.body;
+    const { email = "" } = req.body;
 
     // Validate the phone_number
     let validateError = null;
@@ -144,8 +144,7 @@ export const SendOtp = async (req, res, next) => {
 
     // Send otp
     var otpStatus = null;
-    const name = first_name + " " + last_name;
-    otpStatus = await sendOtp(email, name, otpValue);
+    otpStatus = await sendOtp(email, otpValue);
     if (!otpStatus)
       return ResponseErrorHandler(res, 202, "Failed to send OTP.");
 
@@ -159,6 +158,139 @@ export const SendOtp = async (req, res, next) => {
   } catch (error) {
     console.log("send otp error ->", error);
     return error;
+  }
+};
+
+export const Forgetpassword = async (req, res, next) => {
+  try {
+    const { email = "", otp = "", password = "" } = req.body;
+
+    //check if the user already exists
+    const isUserExists = await DbUser.findOne({
+      $or: [{ email: email }],
+    });
+
+    if (!isUserExists)
+      return ResponseErrorHandler(res, 202, "User not Exists!");
+
+    let otpExpiryLimit = 1;
+    let expiry = new Date();
+    expiry.setMinutes(expiry.getMinutes() - otpExpiryLimit);
+
+    if (otp && password) {
+      const password_hash = await HashPasswords(password);
+
+      let updated = await DbUser.findOneAndUpdate(
+        { email: email },
+        { $set: { password: password_hash } },
+        { upsert: true }
+      );
+      if (!updated)
+        return ResponseErrorHandler(
+          res,
+          202,
+          "Failed to update password. Please contact system admin."
+        );
+
+      let user = { ...updated._doc };
+
+      return ResponseHandler(res, 200, "Password Updated", {
+        isUserUpdated: true,
+      });
+    } else if (otp) {
+      let isOtpExists = await DbOtp.findOne({
+        email: email,
+        createdAt: { $gt: expiry },
+      });
+      if (!isOtpExists)
+        return ResponseErrorHandler(res, 202, "Failed to verify OTP.");
+      else if (isOtpExists) {
+        // Delete the existing otp value
+        await DbOtp.deleteOne({ email: email });
+        return ResponseHandler(res, 200, "Otp Verified", {
+          isOtpVerified: true,
+        });
+      }
+    }
+
+    // Send OTP
+    let isOtpExists = await DbOtp.findOne({
+      email: email,
+      createdAt: { $gt: expiry },
+    });
+    if (isOtpExists)
+      return ResponseErrorHandler(
+        res,
+        202,
+        "Too many OTP requests. Please try after sometime."
+      );
+
+    // Create a unique otp
+    const otpValue = Math.floor(1000 + Math.random() * 9000);
+
+    // Send otp
+    var otpStatus = null;
+    otpStatus = await sendOtp(email, otpValue);
+    if (!otpStatus)
+      return ResponseErrorHandler(res, 202, "Failed to send OTP.");
+
+    const inserted = await DbOtp.create({
+      email,
+      otp: otpValue,
+    });
+    if (!inserted) return ResponseErrorHandler(res, 202, "Failed to send OTP.");
+    return ResponseHandler(res, 200, "Otp Sent", { otp: otpValue });
+  } catch (error) {
+    console.log("send otp error ->", error);
+    return error;
+  }
+};
+
+export const ResendOtp = async (req, res, next) => {
+  try {
+    const { email = "" } = req.body;
+
+    let validateError = null;
+    if (email != "" && !ValidateEmail(email.trim())) {
+      validateError = "Please enter a valid email id i.e abc@domain.com";
+    }
+    if (validateError) return ResponseErrorHandler(res, 202, validateError);
+
+    // Send OTP
+    let otpExpiryLimit = 1;
+    let expiry = new Date();
+    expiry.setMinutes(expiry.getMinutes() - otpExpiryLimit);
+
+    let isOtpExists = await DbOtp.findOne({
+      email: email,
+      createdAt: { $gt: expiry },
+    });
+
+    if (isOtpExists) {
+      return ResponseErrorHandler(res, 202, "To many otp send");
+    }
+
+    // Delete the existing otp value
+    await DbOtp.deleteOne({ email: email });
+
+    // Create a unique otp
+    const otpValue = Math.floor(1000 + Math.random() * 9000);
+
+    // Send otp
+    var otpStatus = null;
+    otpStatus = await sendOtp(email, otpValue);
+    if (!otpStatus)
+      return ResponseErrorHandler(res, 202, "Failed to send OTP.");
+
+    // store the otp on DB and send responce
+    const inserted = await DbOtp.create({
+      email,
+      otp: otpValue,
+    });
+    if (!inserted) return ResponseErrorHandler(res, 202, "Failed to send OTP.");
+    return ResponseHandler(res, 200, "Otp Sent", { otp: otpValue });
+  } catch (err) {
+    HandleServerError(res, req, err);
   }
 };
 
@@ -295,11 +427,11 @@ export const deleteCertificate = async (req, res, next) => {
 /************************************* delete profile function ***************************************/
 export const deleteProfile = async (req, res, next) => {
   try {
-    const { _id } = req.user;
+    const { user_id } = req.body;
     //delete all post of user from post db
-    await DbPost.deleteMany({ owner: _id });
+    await DbPost.deleteMany({ owner: user_id });
 
-    await DbUser.deleteOne({ _id });
+    await DbUser.deleteOne({ _id: user_id });
 
     return ResponseHandler(res, 200, "Profile deleted !");
   } catch (error) {
